matrices: move MatrixBase classes to matrixbase
