+++ doc/src/modules/matrices/matrices.rst
@@ -1,7 +1,7 @@
 Matrices (linear algebra)
 =========================
 
-.. module:: sympy.matrices.matrices
+.. module:: sympy.matrices.matrixbase
 
 Creating Matrices
 -----------------
@@ -531,8 +531,13 @@ Matrix Base Classes
 
 The Matrix classes are built from functionality in various base classes. Every
 methods and attribute of ``Matrix`` is implemented on one of these base
-classes. See also :doc:`/modules/matrices/matrixbase`,
-:doc:`/modules/matrices/dense`, and :doc:`/modules/matrices/sparse`.
+classes. See also :doc:`/modules/matrices/dense`, and
+:doc:`/modules/matrices/sparse`.
+
+.. autoclass:: MatrixCommon
+   :members:
+   :special-members:
+   :inherited-members:
 
 .. autoclass:: MatrixDeterminant
    :members:
