--- doc/src/modules/matrices/matrixbase.rst
@@ -1,9 +1,0 @@
-Matrix base classes
-===================
-
-.. module:: sympy.matrices.matrixbase
-
-.. autoclass:: MatrixCommon
-   :members:
-   :special-members:
-   :inherited-members:
