+++ sympy/matrices/__init__.py
@@ -12,8 +12,7 @@ from .dense import (
     rot_ccw_axis2, rot_ccw_axis3, rot_givens,
     symarray, wronskian, zeros)
 from .dense import MutableDenseMatrix
-from .matrices import MatrixBase
-from .matrixbase import DeferredVector
+from .matrixbase import DeferredVector, MatrixBase
 
 MutableMatrix = MutableDenseMatrix
 Matrix = MutableMatrix
