+++ sympy/matrices/expressions/determinant.py
@@ -3,7 +3,7 @@ from sympy.core.expr import Expr
 from sympy.core.singleton import S
 from sympy.core.sympify import sympify
 from sympy.matrices.exceptions import NonSquareMatrixError
-from sympy.matrices.matrices import MatrixBase
+from sympy.matrices.matrixbase import MatrixBase
 
 
 class Determinant(Expr):
