+++ sympy/matrices/immutable.py
@@ -6,7 +6,7 @@ from sympy.core.cache import cacheit
 from sympy.core.sympify import _sympy_converter as sympify_converter, _sympify
 from sympy.matrices.dense import DenseMatrix
 from sympy.matrices.expressions import MatrixExpr
-from sympy.matrices.matrices import MatrixBase
+from sympy.matrices.matrixbase import MatrixBase
 from sympy.matrices.repmatrix import RepMatrix
 from sympy.matrices.sparse import SparseRepMatrix
 from sympy.multipledispatch import dispatch
