+++ sympy/matrices/sparse.py
@@ -5,7 +5,7 @@ from sympy.utilities.exceptions import sympy_deprecation_warning
 from sympy.utilities.iterables import is_sequence
 from sympy.utilities.misc import as_int
 
-from .matrices import MatrixBase
+from .matrixbase import MatrixBase
 from .repmatrix import MutableRepMatrix, RepMatrix
 
 from .utilities import _iszero
