+++ sympy/matrices/tests/test_subspaces.py
@@ -1,5 +1,5 @@
 from sympy.matrices.common import _MinimalMatrix, _CastableMatrix
-from sympy.matrices.matrices import MatrixSubspaces
+from sympy.matrices.matrixbase import MatrixSubspaces
 from sympy.matrices import Matrix
 from sympy.core.numbers import Rational
 from sympy.core.symbol import symbols
