+++ sympy/physics/quantum/matrixutils.py
@@ -3,7 +3,7 @@
 from sympy.core.expr import Expr
 from sympy.core.numbers import I
 from sympy.core.singleton import S
-from sympy.matrices.matrices import MatrixBase
+from sympy.matrices.matrixbase import MatrixBase
 from sympy.matrices import eye, zeros
 from sympy.external import import_module
