+++ sympy/tensor/array/arrayop.py
@@ -478,7 +478,7 @@ class Flatten(Printable):
     [0, 1, 2, 3, 4, 5]
     """
     def __init__(self, iterable):
-        from sympy.matrices.matrices import MatrixBase
+        from sympy.matrices.matrixbase import MatrixBase
         from sympy.tensor.array import NDimArray
 
         if not isinstance(iterable, (Iterable, MatrixBase)):
@@ -494,7 +494,7 @@ class Flatten(Printable):
         return self
 
     def __next__(self):
-        from sympy.matrices.matrices import MatrixBase
+        from sympy.matrices.matrixbase import MatrixBase
 
         if len(self._iter) > self._idx:
             if isinstance(self._iter, DenseNDimArray):
